Added service and repository test.
