+++ customers-service-reactive/build.gradle
@@ -1,6 +1,6 @@
 plugins {
 	id 'org.springframework.boot' version '2.7.4'
-	id 'io.spring.dependency-management' version '1.0.9.RELEASE'
+	id 'io.spring.dependency-management' version '1.0.13.RELEASE'
 	id 'java'
 	id "io.freefair.lombok" version "6.0.0-m2"
 	id 'jacoco'
