+++ customers-service-reactive/src/test/java/com/petclinic/customersservice/presentationlayer/PhotoControllerUnitTest.java
@@ -23,30 +23,36 @@ class PhotoControllerUnitTest {
 
     private final String PHOTO_ID = photo.getId();
 
+    private final String PHOTO_ID_NOT_FOUND = "Not found";
+
     @Autowired
     private WebTestClient client;
 
     @MockBean
     PhotoService photoService;
 
-    @Test
-    public void getPhotoByPhotoId() {
 
-        when(photoService.getPhotoByPhotoId(anyString())).thenReturn(Mono.just(photo));
 
-        client.get()
-                .uri("/photos/" + PHOTO_ID)
-                .accept(MediaType.APPLICATION_JSON)
-                .exchange()
-                .expectStatus().isOk()
-                .expectHeader().contentType(MediaType.APPLICATION_JSON)
-                .expectBody()
-                .jsonPath("$.name").isEqualTo(photo.getName())
-                .jsonPath("$.type").isEqualTo(photo.getType())
-                .jsonPath("$.photo").isEqualTo(photo.getPhoto());
+//    @Test
+//    public void getPhotoByPhotoIdNotFound() {
+//
+//        when(photoService.getPhotoByPhotoId(anyString())).thenReturn(Mono.just(photo));
+//
+//        client.get()
+//                .uri("/photos/" + PHOTO_ID_NOT_FOUND)
+//                .accept(MediaType.APPLICATION_JSON)
+//                .exchange()
+//                .expectStatus().isNotFound();
+//                .expectHeader().contentType(MediaType.APPLICATION_JSON)
+//                .expectBody()
+//                .jsonPath("$.name").isEqualTo(photo.getName())
+//                .jsonPath("$.type").isEqualTo(photo.getType())
+//                .jsonPath("$.photo").isEqualTo(photo.getPhoto());
+//
+//        Mockito.verify(photoService, times(1)).getPhotoByPhotoId(PHOTO_ID);
+//    }
 
-        Mockito.verify(photoService, times(1)).getPhotoByPhotoId(PHOTO_ID);
-    }
+
 
 
     private Photo buildPhoto() {
